Data.Functor.Infix

Documentation

(<$$>) :: forall a b f f. (Functor f, Functor f) => (a -> b) -> f (f a) -> f (f b) infixl 4

(<$$$>) :: forall a b f f f. (Functor f, Functor f, Functor f) => (a -> b) -> f (f (f a)) -> f (f (f b)) infixl 4

(<$$$$>) :: forall a b f f f f. (Functor f, Functor f, Functor f, Functor f) => (a -> b) -> f (f (f (f a))) -> f (f (f (f b))) infixl 4

(<$$$$$>) :: forall a b f f f f f. (Functor f, Functor f, Functor f, Functor f, Functor f) => (a -> b) -> f (f (f (f (f a)))) -> f (f (f (f (f b)))) infixl 4

(<$$$$$$>) :: forall a b f f f f f f. (Functor f, Functor f, Functor f, Functor f, Functor f, Functor f) => (a -> b) -> f (f (f (f (f (f a))))) -> f (f (f (f (f (f b))))) infixl 4

(<$$$$$$$>) :: forall a b f f f f f f f. (Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f) => (a -> b) -> f (f (f (f (f (f (f a)))))) -> f (f (f (f (f (f (f b)))))) infixl 4

(<$$$$$$$$>) :: forall a b f f f f f f f f. (Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f) => (a -> b) -> f (f (f (f (f (f (f (f a))))))) -> f (f (f (f (f (f (f (f b))))))) infixl 4

(<$$$$$$$$$>) :: forall a b f f f f f f f f f. (Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f) => (a -> b) -> f (f (f (f (f (f (f (f (f a)))))))) -> f (f (f (f (f (f (f (f (f b)))))))) infixl 4

(<$$$$$$$$$$>) :: forall a b f f f f f f f f f f. (Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f) => (a -> b) -> f (f (f (f (f (f (f (f (f (f a))))))))) -> f (f (f (f (f (f (f (f (f (f b))))))))) infixl 4

(<$$$$$$$$$$$>) :: forall a b f f f f f f f f f f f. (Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f) => (a -> b) -> f (f (f (f (f (f (f (f (f (f (f a)))))))))) -> f (f (f (f (f (f (f (f (f (f (f b)))))))))) infixl 4

(<$$$$$$$$$$$$>) :: forall a b f f f f f f f f f f f f. (Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f) => (a -> b) -> f (f (f (f (f (f (f (f (f (f (f (f a))))))))))) -> f (f (f (f (f (f (f (f (f (f (f (f b))))))))))) infixl 4

(<$$$$$$$$$$$$$>) :: forall a b f f f f f f f f f f f f f. (Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f) => (a -> b) -> f (f (f (f (f (f (f (f (f (f (f (f (f a)))))))))))) -> f (f (f (f (f (f (f (f (f (f (f (f (f b)))))))))))) infixl 4

(<$$$$$$$$$$$$$$>) :: forall a b f f f f f f f f f f f f f f. (Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f) => (a -> b) -> f (f (f (f (f (f (f (f (f (f (f (f (f (f a))))))))))))) -> f (f (f (f (f (f (f (f (f (f (f (f (f (f b))))))))))))) infixl 4

(<$$$$$$$$$$$$$$$>) :: forall a b f f f f f f f f f f f f f f f. (Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f) => (a -> b) -> f (f (f (f (f (f (f (f (f (f (f (f (f (f (f a)))))))))))))) -> f (f (f (f (f (f (f (f (f (f (f (f (f (f (f b)))))))))))))) infixl 4

(<$$$$$$$$$$$$$$$$>) :: forall a b f f f f f f f f f f f f f f f f. (Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f) => (a -> b) -> f (f (f (f (f (f (f (f (f (f (f (f (f (f (f (f a))))))))))))))) -> f (f (f (f (f (f (f (f (f (f (f (f (f (f (f (f b))))))))))))))) infixl 4

(<$$$$$$$$$$$$$$$$$>) :: forall a b f f f f f f f f f f f f f f f f f. (Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f) => (a -> b) -> f (f (f (f (f (f (f (f (f (f (f (f (f (f (f (f (f a)))))))))))))))) -> f (f (f (f (f (f (f (f (f (f (f (f (f (f (f (f (f b)))))))))))))))) infixl 4

(<$$$$$$$$$$$$$$$$$$>) :: forall a b f f f f f f f f f f f f f f f f f f. (Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f) => (a -> b) -> f (f (f (f (f (f (f (f (f (f (f (f (f (f (f (f (f (f a))))))))))))))))) -> f (f (f (f (f (f (f (f (f (f (f (f (f (f (f (f (f (f b))))))))))))))))) infixl 4

(<$$$$$$$$$$$$$$$$$$$>) :: forall a b f f f f f f f f f f f f f f f f f f f. (Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f) => (a -> b) -> f (f (f (f (f (f (f (f (f (f (f (f (f (f (f (f (f (f (f a)))))))))))))))))) -> f (f (f (f (f (f (f (f (f (f (f (f (f (f (f (f (f (f (f b)))))))))))))))))) infixl 4

(<$$$$$$$$$$$$$$$$$$$$>) :: forall a b f f f f f f f f f f f f f f f f f f f f. (Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f) => (a -> b) -> f (f (f (f (f (f (f (f (f (f (f (f (f (f (f (f (f (f (f (f a))))))))))))))))))) -> f (f (f (f (f (f (f (f (f (f (f (f (f (f (f (f (f (f (f (f b))))))))))))))))))) infixl 4

(<&>) :: forall a b f. Functor f => f a -> (a -> b) -> f b infixl 1

(<&&>) :: forall a b f f. (Functor f, Functor f) => f (f a) -> (a -> b) -> f (f b) infixl 1

(<&&&>) :: forall a b f f f. (Functor f, Functor f, Functor f) => f (f (f a)) -> (a -> b) -> f (f (f b)) infixl 1

(<&&&&>) :: forall a b f f f f. (Functor f, Functor f, Functor f, Functor f) => f (f (f (f a))) -> (a -> b) -> f (f (f (f b))) infixl 1

(<&&&&&>) :: forall a b f f f f f. (Functor f, Functor f, Functor f, Functor f, Functor f) => f (f (f (f (f a)))) -> (a -> b) -> f (f (f (f (f b)))) infixl 1

(<&&&&&&>) :: forall a b f f f f f f. (Functor f, Functor f, Functor f, Functor f, Functor f, Functor f) => f (f (f (f (f (f a))))) -> (a -> b) -> f (f (f (f (f (f b))))) infixl 1

(<&&&&&&&>) :: forall a b f f f f f f f. (Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f) => f (f (f (f (f (f (f a)))))) -> (a -> b) -> f (f (f (f (f (f (f b)))))) infixl 1

(<&&&&&&&&>) :: forall a b f f f f f f f f. (Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f) => f (f (f (f (f (f (f (f a))))))) -> (a -> b) -> f (f (f (f (f (f (f (f b))))))) infixl 1

(<&&&&&&&&&>) :: forall a b f f f f f f f f f. (Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f) => f (f (f (f (f (f (f (f (f a)))))))) -> (a -> b) -> f (f (f (f (f (f (f (f (f b)))))))) infixl 1

(<&&&&&&&&&&>) :: forall a b f f f f f f f f f f. (Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f) => f (f (f (f (f (f (f (f (f (f a))))))))) -> (a -> b) -> f (f (f (f (f (f (f (f (f (f b))))))))) infixl 1

(<&&&&&&&&&&&>) :: forall a b f f f f f f f f f f f. (Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f) => f (f (f (f (f (f (f (f (f (f (f a)))))))))) -> (a -> b) -> f (f (f (f (f (f (f (f (f (f (f b)))))))))) infixl 1

(<&&&&&&&&&&&&>) :: forall a b f f f f f f f f f f f f. (Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f) => f (f (f (f (f (f (f (f (f (f (f (f a))))))))))) -> (a -> b) -> f (f (f (f (f (f (f (f (f (f (f (f b))))))))))) infixl 1

(<&&&&&&&&&&&&&>) :: forall a b f f f f f f f f f f f f f. (Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f) => f (f (f (f (f (f (f (f (f (f (f (f (f a)))))))))))) -> (a -> b) -> f (f (f (f (f (f (f (f (f (f (f (f (f b)))))))))))) infixl 1

(<&&&&&&&&&&&&&&>) :: forall a b f f f f f f f f f f f f f f. (Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f) => f (f (f (f (f (f (f (f (f (f (f (f (f (f a))))))))))))) -> (a -> b) -> f (f (f (f (f (f (f (f (f (f (f (f (f (f b))))))))))))) infixl 1

(<&&&&&&&&&&&&&&&>) :: forall a b f f f f f f f f f f f f f f f. (Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f) => f (f (f (f (f (f (f (f (f (f (f (f (f (f (f a)))))))))))))) -> (a -> b) -> f (f (f (f (f (f (f (f (f (f (f (f (f (f (f b)))))))))))))) infixl 1

(<&&&&&&&&&&&&&&&&>) :: forall a b f f f f f f f f f f f f f f f f. (Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f) => f (f (f (f (f (f (f (f (f (f (f (f (f (f (f (f a))))))))))))))) -> (a -> b) -> f (f (f (f (f (f (f (f (f (f (f (f (f (f (f (f b))))))))))))))) infixl 1

(<&&&&&&&&&&&&&&&&&>) :: forall a b f f f f f f f f f f f f f f f f f. (Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f) => f (f (f (f (f (f (f (f (f (f (f (f (f (f (f (f (f a)))))))))))))))) -> (a -> b) -> f (f (f (f (f (f (f (f (f (f (f (f (f (f (f (f (f b)))))))))))))))) infixl 1

(<&&&&&&&&&&&&&&&&&&>) :: forall a b f f f f f f f f f f f f f f f f f f. (Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f) => f (f (f (f (f (f (f (f (f (f (f (f (f (f (f (f (f (f a))))))))))))))))) -> (a -> b) -> f (f (f (f (f (f (f (f (f (f (f (f (f (f (f (f (f (f b))))))))))))))))) infixl 1

(<&&&&&&&&&&&&&&&&&&&>) :: forall a b f f f f f f f f f f f f f f f f f f f. (Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f) => f (f (f (f (f (f (f (f (f (f (f (f (f (f (f (f (f (f (f a)))))))))))))))))) -> (a -> b) -> f (f (f (f (f (f (f (f (f (f (f (f (f (f (f (f (f (f (f b)))))))))))))))))) infixl 1

(<&&&&&&&&&&&&&&&&&&&&>) :: forall a b f f f f f f f f f f f f f f f f f f f f. (Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f, Functor f) => f (f (f (f (f (f (f (f (f (f (f (f (f (f (f (f (f (f (f (f a))))))))))))))))))) -> (a -> b) -> f (f (f (f (f (f (f (f (f (f (f (f (f (f (f (f (f (f (f (f b))))))))))))))))))) infixl 1

(<$>) :: Functor f => (a -> b) -> f a -> f b infixl 4

An infix synonym for fmap.

The name of this operator is an allusion to $. Note the similarities between their types:

 ($)  ::              (a -> b) ->   a ->   b
(<$>) :: Functor f => (a -> b) -> f a -> f b

Whereas $ is function application, <$> is function application lifted over a Functor.

Examples

Expand

Convert from a Maybe Int to a Maybe String using show:

>>> show <$> Nothing
Nothing
>>> show <$> Just 3
Just "3"

Convert from an Either Int Int to an Either Int String using show:

>>> show <$> Left 17
Left 17
>>> show <$> Right 17
Right "17"

Double each element of a list:

>>> (*2) <$> [1,2,3]
[2,4,6]

Apply even to the second element of a pair:

>>> even <$> (2,2)
(2,True)

declareInfixFmapForFunctorCompositionOfDegree :: Int -> Q [Dec]

declareFlippedInfixFmapForFunctorCompositionOfDegree :: Int -> Q [Dec]

declareInfixFmapN :: Int -> Q [Dec]

Deprecated: Use declareInfixFmapForFunctorCompositionOfDegree and/or reconsider your life choices.

declareInfixPamfN :: Int -> Q [Dec]

Deprecated: Use declareFlippedInfixFmapForFunctorCompositionOfDegree and/or reconsider your life choices.